src/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
azure-core==1.32.0
azure-cosmos==4.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.5
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
multidict==6.1.0
packaging==24.2
propcache==0.2.0
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.17.2
setuptools==16.0
